requirements.txt
pyopengl
pillow
